Teste do CRUD do User e Criação de JSON dinâmico + pequenos ajustes

+++ __tests__/unit/test_calculadora.py
@@ -74,6 +74,6 @@ def test_somar_dois_numeros_param(num1, num2, resultado_esperado):
                          ler_csv('./fixtures/csv/massaSomar.csv'))
 def test_somar_dois_numeros_csv(num1, num2, resultado_esperado):
     
-    resultado_obtido = somar_dois_numeros(int(num1), int(num2))
-    assert int(resultado_esperado) == resultado_obtido
+    resultado_obtido = somar_dois_numeros(float(num1), float(num2))
+    assert float(resultado_esperado) == resultado_obtido
